+++ src/Infrastructure/Condition/Evaluator/Creator/Chain.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator;
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator;
 
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionEvaluator;
