+++ src/Infrastructure/Condition/Evaluator/Creator/CompositeEvaluatorCreator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator;
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator;
 
 use ArtARTs36\MergeRequestLinter\Application\Condition\Evaluators\Composite\CompositeEvaluator;
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionEvaluator;
