+++ src/Infrastructure/Condition/Evaluator/Creator/ContainsHeadingEvaluatorCreator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator;
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator;
 
 use ArtARTs36\MergeRequestLinter\Application\Condition\Evaluators\Strings\Markdown\ContainsHeadingEvaluator;
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionEvaluator;
