+++ src/Infrastructure/Condition/EvaluatorFactory.php
@@ -2,8 +2,8 @@
 
 namespace ArtARTs36\MergeRequestLinter\Infrastructure\Condition;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\EvaluatorCreator;
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionEvaluator;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\EvaluatorCreator;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Exceptions\ConditionEvaluatorNotFound;
 
 /**
