+++ tests/Mocks/MockEvaluatorCreator.php
@@ -3,7 +3,7 @@
 namespace ArtARTs36\MergeRequestLinter\Tests\Mocks;
 
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionEvaluator;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\EvaluatorCreator;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\EvaluatorCreator;
 
 final class MockEvaluatorCreator implements EvaluatorCreator
 {
