+++ tests/Unit/Infrastructure/Condition/Evaluator/Creator/ChainFactoryTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Condition\Evaluator;
+namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Condition\Evaluator\Creator;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Chain;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\ChainFactory;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\Chain;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\ChainFactory;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\SubjectFactory;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\ArrayMap;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockPropertyExtractor;
@@ -12,8 +12,8 @@ use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 final class ChainFactoryTest extends TestCase
 {
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\ChainFactory::create
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\ChainFactory::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\ChainFactory::create
+     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\ChainFactory::__construct
      */
     public function testCreate(): void
     {
