+++ tests/Unit/Infrastructure/Condition/Evaluator/Creator/ChainTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Condition\Evaluator;
+namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Condition\Evaluator\Creator;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Chain;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\Chain;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockConditionEvaluator;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockEvaluatorCreator;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
@@ -10,9 +10,9 @@ use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 final class ChainTest extends TestCase
 {
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Chain::create
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Chain::add
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Chain::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\Chain::create
+     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\Chain::add
+     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Condition\Evaluator\Creator\Chain::__construct
      */
     public function testCreate(): void
     {
